refactor : Modify name of Map component
