+++ src/App.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { Route, Routes } from 'react-router-dom'
 
-import Map from '@pages/Map/Map'
+import Main from '@pages/Main'
 
 function App() {
   return (
     <div>
       <Routes>
-        <Route path="/" element={<Map />} />
+        <Route path="/" element={<Main />} />
       </Routes>
     </div>
   )
